added model

+++ EmailTemplateteService/TemplateParams.cs
@@ -2,7 +2,7 @@
 {
     public class TemplateParams
     {
-        public int? BranchId { get; set; }
+        public short? BranchId { get; set; }
         public int? CampaignId { get; set; }
         public string Subject { get; set; }
         public string Body { get; set; }
